Created a routes folder and put everything related to the page there, removed unnecessary things from App.css
