+++ src/components/Search/Search.module.css
@@ -1,22 +1,3 @@
-.character_link {
-    display: inline-block;
-    margin-right: 10px;
-    color: black;
-}
-
-.character_link:hover {
-    text-decoration: none;
-}
-
-h1 {
-    margin-left: 50px;
-}
-
-.charactersCount {
-    font-size: 32px;
-    color: grey;
-}
-
 .search_container {
     display: flex;
     justify-content: space-between;
