+++ src/routes/Characters/CharacterDetails.module.css
