+++ src/routes/Comics/Comics.module.css
@@ -23,21 +23,3 @@ h1 {
     font-size: 32px;
     color: grey;
 }
-
-.search_container {
-    display: flex;
-    justify-content: space-between;
-    margin: 10px 50px 0;
-}
-
-.search_input {
-    width: calc(100% - 120px);
-    margin-right: 20px;
-}
-
-.search_button {
-    width: 250px;
-    background-color: orange;
-    color: white;
-    border-radius: 0;
-}
